git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/array-rework@3172 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libFreeMat/demo/Global.hpp
@@ -3,7 +3,8 @@
 
 #include <QtGlobal>
 
-typedef double index_t;
+//typedef double index_t;
+typedef int index_t;
 const int NDims = 6;
 typedef QString Exception;
